% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AminoAcidToCodon.R
\name{SequenceAnalysis.AminoAcidToCodon}
\alias{SequenceAnalysis.AminoAcidToCodon}
\title{SequenceAnalysis.AminoAcidToCodon}
\usage{
SequenceAnalysis.AminoAcidToCodon(AminoAcid = NULL)
}
\arguments{
\item{AminoAcid}{Amino Acid}
}
\value{
Related Codons
}
\description{
All related codons of desired amino acid.
}
\details{
All related codons of desired amino acid.
}
\examples{
SequenceAnalysis.AminoAcidToCodon(AminoAcid="A")
}
\author{
Babak Khorsand
}

