% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonToAminoAcid.R
\name{SequenceAnalysis.CodonToAminoAcid}
\alias{SequenceAnalysis.CodonToAminoAcid}
\title{SequenceAnalysis.CodonToAminoAcid}
\usage{
SequenceAnalysis.CodonToAminoAcid(Nucleotide_Sequence = NULL)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
Related amino acid
}
\description{
Converting desired codons to related amino acids.
}
\details{
Converting nucleotide sequence to protein sequence.
}
\examples{
SequenceAnalysis.CodonToAminoAcid(Nucleotide_Sequence="atggctgcagcggccagtcacgatcagaggtaagttgtc")
}
\author{
Babak Khorsand
}

