% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonUsage.R
\name{SequenceAnalysis.CodonUsage}
\alias{SequenceAnalysis.CodonUsage}
\title{SequenceAnalysis.CodonUsage}
\usage{
SequenceAnalysis.CodonUsage(Nucleotide_Sequence = NULL, UniprotKB = NULL)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}

\item{UniprotKB}{UniProt ID of desired sequence}
}
\value{
CodonUsage
}
\description{
Frequency of occurrence of synonymous codons
}
\details{
Codon Usage will be calculated for input sequence or by passing UniProt ID it will catch the sequence online from UniProt database and then calculate Codon Usage for it.
}
\examples{
SequenceAnalysis.CodonUsage(Nucleotide_Sequence="atggctgcagcggccagtcacgatcagaggtaagttgtc")
SequenceAnalysis.CodonUsage(UniprotKB="O15131")
}
\author{
Babak Khorsand
}

