% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complement.R
\name{SequenceAnalysis.Complement}
\alias{SequenceAnalysis.Complement}
\title{SequenceAnalysis.Complement}
\usage{
SequenceAnalysis.Complement(Nucleotide_Sequence = NULL, UniprotKB = NULL,
  CDS = FALSE)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}

\item{UniprotKB}{UniProt ID of desired sequence}

\item{CDS}{if TRUE Stacking Energy of CDS Region will be calculated}
}
\value{
Complement
}
\description{
Complement of desired nucleotide sequence.
}
\details{
Complement of desired nucleotide sequence. (A for T, C for G and vice versa) (tgacta -> Complement: actgat)
}
\examples{
SequenceAnalysis.Complement("actagtcacgatcag")
SequenceAnalysis.Complement(UniprotKB="O15131")
SequenceAnalysis.Complement(UniprotKB="O15131",CDS=TRUE)
}
\author{
Babak Khorsand
}

