% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC.R
\name{SequenceAnalysis.GC}
\alias{SequenceAnalysis.GC}
\title{SequenceAnalysis.GC}
\usage{
SequenceAnalysis.GC(Nucleotide_Sequence = NULL, UniprotKB = NULL,
  CDS = FALSE)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}

\item{UniprotKB}{UniProt ID of desired sequence}

\item{CDS}{if TRUE GC content of CDS Region will be calculated}
}
\value{
GC content
}
\description{
GC content : percentage of nucleotide g and c
}
\details{
Calculate GC content for input sequence or by passing UniProt ID it will catch the sequence online from UniProt database and then calculate GC content for it.
}
\examples{
SequenceAnalysis.GC("actagtcacgatcag")
SequenceAnalysis.GC(UniprotKB="O15131")
SequenceAnalysis.GC(UniprotKB="O15131",CDS=TRUE)
}
\author{
Babak Khorsand
}

