% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLocalization.R
\name{SequenceAnalysis.GetLocalization}
\alias{SequenceAnalysis.GetLocalization}
\title{SequenceAnalysis.GetLocalization}
\usage{
SequenceAnalysis.GetLocalization(UniprotKB)
}
\arguments{
\item{UniprotKB}{Uniprot ID of a sequence}
}
\value{
Subcellular location of that protein
}
\description{
Subcellular location of that protein
}
\details{
Getting Localization From Gene Ontology Cellular Component inside Uniprot.
}
\examples{
SequenceAnalysis.GetLocalization("O15131")
}
\author{
Babak Khorsand
}

