% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProteinOrganism.R
\name{SequenceAnalysis.GetProteinOrganism}
\alias{SequenceAnalysis.GetProteinOrganism}
\title{SequenceAnalysis.GetProteinOrganism}
\usage{
SequenceAnalysis.GetProteinOrganism(UniprotKB)
}
\arguments{
\item{UniprotKB}{UniProt ID of a sequence}
}
\value{
Protein, Gene and Organism of desired UniProt ID.
}
\description{
Get Protein Organism
}
\details{
Get Protein, Gene and Organism of desired protein from UniProt
}
\examples{
SequenceAnalysis.GetProteinOrganism("P03485")
}
\author{
Babak Khorsand
}

