% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProteinSequence.R
\name{SequenceAnalysis.GetProteinSequence}
\alias{SequenceAnalysis.GetProteinSequence}
\title{SequenceAnalysis.GetProteinSequence}
\usage{
SequenceAnalysis.GetProteinSequence(UniprotKB)
}
\arguments{
\item{UniprotKB}{UniProt ID of a sequence}
}
\value{
Protein Sequence
}
\description{
Get Protein Sequence From UniProt
}
\details{
Get Protein Sequence From UniProt by UniProt ID
}
\examples{
SequenceAnalysis.GetProteinSequence("O15131")
}
\author{
Babak Khorsand
}

