% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reverse.R
\name{SequenceAnalysis.Reverse}
\alias{SequenceAnalysis.Reverse}
\title{SequenceAnalysis.Reverse}
\usage{
SequenceAnalysis.Reverse(Nucleotide_Sequence = NULL, UniprotKB = NULL,
  CDS = FALSE)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}

\item{UniprotKB}{UniProt ID of desired sequence}

\item{CDS}{if TRUE Reverse of CDS Region will be calculated}
}
\value{
Reverse
}
\description{
Reverse of desired nucleotide sequence.
}
\details{
Reverse of desired nucleotide sequence. (atcagt -> Reverse: tgacta)
}
\examples{
SequenceAnalysis.Reverse("actagtcacgatcag")
SequenceAnalysis.Reverse(UniprotKB="O15131")
SequenceAnalysis.Reverse(UniprotKB="O15131",CDS=TRUE)
}
\author{
Babak Khorsand
}

