% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNuclotideSequence.R
\name{SequenceAnalysis.GetNucleotideSequence}
\alias{SequenceAnalysis.GetNucleotideSequence}
\title{SequenceAnalysis.GetNucleotideSequence}
\usage{
SequenceAnalysis.GetNucleotideSequence(UniprotKB)
}
\arguments{
\item{UniprotKB}{UniProt ID of a sequence}
}
\value{
List The first element specify UniprotKB, The second element specify nucleotide sequence of coding region and the third element return whole nucleotide sequence.
}
\description{
Get Nucleotide Sequence
}
\details{
Get Nucleotide Sequence from EBI database by UniProt ID
}
\examples{
SequenceAnalysis.GetNucleotideSequence("O15131")
}
\author{
Babak Khorsand
}

